Removed the no longer needed Geographic Identity Classes and Objects
