+++ src/corelib/corelib.csproj
@@ -90,6 +90,7 @@
     <Compile Include="Core\ResponseExtensions.cs" />
     <Compile Include="Core\RestWebHeaderCollection.cs" />
     <Compile Include="Core\Validators\NamespaceDoc.cs" />
+    <Compile Include="Providers\Rackspace\CloudIdentityProvider.cs" />
     <Compile Include="Providers\Rackspace\Exceptions\BulkDeletionException.cs" />
     <Compile Include="Core\HttpStatusCodeParser.cs" />
     <Compile Include="Core\IStatusParser.cs" />
@@ -231,15 +232,12 @@
     <Compile Include="Core\Caching\UserAccessCache.cs" />
     <Compile Include="Core\Providers\IIdentityProvider.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Providers\Rackspace\GeographicalCloudIdentityProvider.cs" />
-    <Compile Include="Providers\Rackspace\CloudIdentityProviderFactory.cs" />
     <Compile Include="Core\Domain\UserAccess.cs" />
     <Compile Include="Providers\Rackspace\Objects\Request\AuthRequest.cs" />
     <Compile Include="Core\Domain\Role.cs" />
     <Compile Include="Core\Domain\Endpoint.cs" />
     <Compile Include="Core\Domain\ServiceCatalog.cs" />
     <Compile Include="Providers\Rackspace\Objects\Response\AuthenticationResponse.cs" />
-    <Compile Include="Providers\Rackspace\CloudIdentityProvider.cs" />
     <Compile Include="Core\Domain\UserDetails.cs" />
     <Compile Include="Providers\Rackspace\Objects\Response\RolesResponse.cs" />
     <Compile Include="Providers\Rackspace\Objects\Request\UpdateServerRequest.cs" />
