+++ src/testing/integration/Providers/Rackspace/ExtendedIdentityTests.cs
@@ -52,7 +52,7 @@ namespace Net.OpenStack.Testing.Integration.Providers.Rackspace
 
         private static IExtendedCloudIdentityProvider BuildProvider(IRestService restService = null, ICache<UserAccess> cache = null)
         {
-            return new CloudIdentityProvider(restService, cache, Bootstrapper.Settings.RackspaceExtendedIdentityUSUrl, Bootstrapper.Settings.RackspaceExtendedIdentityUKUrl);
+            return new CloudIdentityProvider(restService, cache, Bootstrapper.Settings.RackspaceExtendedIdentityUrl);
         }
 
         [TestMethod]
@@ -108,8 +108,7 @@ namespace Net.OpenStack.Testing.Integration.Providers.Rackspace
                 provider.Authenticate(new RackspaceCloudIdentity
                 {
                     Username = _testAdminIdentity.Username,
-                    Password = AdminNewPassword,
-                    CloudInstance = _testAdminIdentity.CloudInstance
+                    Password = AdminNewPassword
                 });
 
             Assert.IsNotNull(userAcess);
@@ -155,8 +154,7 @@ namespace Net.OpenStack.Testing.Integration.Providers.Rackspace
                 provider.Authenticate(new RackspaceCloudIdentity
                 {
                     Username = _testAdminIdentity.Username,
-                    Password = AdminNewPassword,
-                    CloudInstance = _testAdminIdentity.CloudInstance
+                    Password = AdminNewPassword
                 });
 
             Assert.IsNotNull(userAcess);
